Update Open discussion modal and refactor submit button logic
