+++ apps/financiers/financiers/src/app/marketplace/title/modal/modal.component.scss
@@ -1,0 +1,4 @@
+mat-spinner {
+  margin-left: 8px;
+  display: inline-block;
+}
