+++ apps/financiers/financiers/src/app/marketplace/title/view/view.component.ts
@@ -1,32 +1,15 @@
-﻿import { Component, ChangeDetectionStrategy, ViewChild, TemplateRef } from '@angular/core';
-import { getCurrencySymbol } from '@angular/common';
+﻿import { ChangeDetectionStrategy, Component, TemplateRef, ViewChild } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { ActivatedRoute, Router } from '@angular/router';
-import { mainRoute, additionalRoute, artisticRoute, productionRoute } from '@blockframes/movie/marketplace';
-import { Campaign, Movie, Organization } from '@blockframes/model';
-import { OrganizationService } from '@blockframes/organization/+state';
-import { CampaignService } from '@blockframes/campaign/+state';
-import { RouteDescription } from '@blockframes/utils/common-interfaces';
-import { SendgridService } from '@blockframes/utils/emails/sendgrid.service';
-import { templateIds } from '@blockframes/utils/emails/ids';
 import { MatDialog } from '@angular/material/dialog';
-import { pluck, shareReplay, switchMap, tap } from 'rxjs/operators';
-import { AuthService } from '@blockframes/auth/+state';
-import { UserService } from '@blockframes/user/+state';
-import { ErrorResultResponse } from '@blockframes/utils/utils';
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { getUserEmailData, OrgEmailData } from '@blockframes/utils/emails/utils';
+import { ActivatedRoute, Router } from '@angular/router';
+import { CampaignService } from '@blockframes/campaign/+state';
+import { Movie, Organization } from '@blockframes/model';
 import { MovieService } from '@blockframes/movie/+state/movie.service';
-import { supportMailosaur } from '@blockframes/utils/constants';
-import { SnackbarErrorComponent } from '@blockframes/ui/snackbar/error/snackbar-error.component';
+import { additionalRoute, artisticRoute, mainRoute, productionRoute } from '@blockframes/movie/marketplace';
+import { OrganizationService } from '@blockframes/organization/+state';
+import { RouteDescription } from '@blockframes/utils/common-interfaces';
+import { pluck, shareReplay, switchMap, tap } from 'rxjs/operators';
 import { MarketplaceMovieModalComponent } from '../modal/modal.component';
-
-interface EmailData {
-  subject: string;
-  from: string;
-  to: string;
-  message: string;
-}
 
 @Component({
   selector: 'financiers-movie-view',
@@ -48,7 +31,8 @@ export class MarketplaceMovieViewComponent {
   );
 
   public campaign$ = this.movie$.pipe(
-    switchMap(movie => this.campaignService.valueChanges(movie.id))
+    switchMap(movie => this.campaignService.valueChanges(movie.id)),
+    tap(campaign => this.currency = campaign.currency)
   );
 
   public currency: string;
@@ -79,13 +63,9 @@ export class MarketplaceMovieViewComponent {
   constructor(
     private route: ActivatedRoute,
     private movieService: MovieService,
-    private authService: AuthService,
     private orgService: OrganizationService,
-    private userService: UserService,
     private campaignService: CampaignService,
     private dialog: MatDialog,
-    private snackBar: MatSnackBar,
-    private sendgrid: SendgridService,
     public router: Router
   ) { }
 
@@ -103,82 +83,16 @@ export class MarketplaceMovieViewComponent {
         orgs,
         form,
         movie,
+        currency: this.currency,
         campaign: this.movie$.pipe(
           switchMap(movie => this.campaignService.valueChanges(movie.id))
         ),
-        onSend: this.sendEmail
+        closeModal: this.close
       }
     });
   }
 
-  async sendEmail(emailData: EmailData, title: string, orgs: Organization[]) {
-    try {
-      const templateId = templateIds.financiers.invest;
-      const userSubject = getUserEmailData(this.authService.profile);
-
-      const org = this.orgService.org;
-      const orgUserSubject: OrgEmailData = {
-        denomination: org.denomination.full ?? org.denomination.public,
-        id: org.id || '',
-        email: org.email || ''
-      }
-
-      const promises: Promise<ErrorResultResponse>[] = [];
-
-      const cyCheck = 'Cypress' in window;
-      let emailReady = false;
-      let numEmails = 0;
-
-      for (const org of orgs) {
-        const users = await this.userService.getValue(org.userIds);
-        for (const user of users) {
-          let toUser = getUserEmailData(user);
-          const userEmail = toUser.email;
-          // For e2e test purpose
-          if (cyCheck) {
-            toUser = { ...toUser, email: supportMailosaur };
-          }
-
-          const data = {
-            ...emailData,
-            currency: getCurrencySymbol(this.currency, 'wide'),
-            userSubject,
-            user: toUser,
-            org: orgUserSubject,
-            title,
-          };
-
-          /*
-           * If running E2E, for user other than sender,
-           * store it for access in E2E test.
-           * A single email is sufficient to check the email template
-           */
-          if (cyCheck && (userEmail !== userSubject.email)) {
-            emailReady = true;
-            ++numEmails;
-            window['cyEmailData'] = data;
-          }
-
-          const promise = this.sendgrid.sendWithTemplate({
-            request: { templateId, data, to: toUser.email },
-            app: 'financiers'
-          });
-          promises.push(promise);
-        }
-      }
-
-      if (cyCheck && emailReady) {
-        window['cyEmailData'].numEmails = numEmails;
-      }
-      const res = await Promise.all(promises);
-      const success = res.some(r => r.result);      
-      if (success) {
-        this.snackBar.open('Your email has been sent.', null, { duration: 3000 });
-      } else {
-        throw new Error('An error occured');
-       }
-    } catch (err) {
-      this.snackBar.openFromComponent(SnackbarErrorComponent, { duration: 5000 });
-    }
+  close = () => {
+    this.dialog.closeAll();
   }
 }
